requirements.txt
pandas
SimpleITK
numpy
matplotlib
pathlib
monai
pyyaml
pyradiomics